requirements.txt
# oauth2client
# gspread
# flask
# jinja2
# requests
# fake-useragent
# psycopg2
# config
google-api-python-client
google-auth-httplib2
google-auth-oauthlib